chore: moved test funcs and data to testdata file
