+++ tests/e2e_testdata.go
@@ -1,6 +1,54 @@
 package tests
 
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"net/http"
+	"time"
+
+	"github.com/adshmh/meter/db"
+	"github.com/gojektech/heimdall/httpclient"
+	influxdb2 "github.com/influxdata/influxdb-client-go/v2"
+	influxAPI "github.com/influxdata/influxdb-client-go/v2/api"
+	influxDomain "github.com/influxdata/influxdb-client-go/v2/domain"
+	"github.com/pokt-foundation/portal-api-go/repository"
+	stringUtils "github.com/pokt-foundation/utils-go/strings"
+	"github.com/stretchr/testify/require"
+)
+
+// TODO move this to utils-go
+func startOfDay(day time.Time) time.Time {
+	y, m, d := day.Date()
+	l := day.Location()
+
+	return time.Date(y, m, d, 0, 0, 0, 0, l)
+}
+
 var (
+	// Test vars
+	ctx               = context.Background()
+	today             = startOfDay(time.Now().UTC())
+	endOfDay          = today.AddDate(0, 0, 1)
+	todayFormatted    = today.Format(time.RFC3339)
+	endOfDayFormatted = endOfDay.Format(time.RFC3339)
+	influxOptions     = db.InfluxDBOptions{
+		URL:                 "http://localhost:8086",
+		Token:               "mytoken",
+		Org:                 "myorg",
+		DailyBucket:         "mainnetRelayApp1d",
+		CurrentBucket:       "mainnetRelayApp10m",
+		DailyOriginBucket:   "mainnetOrigin1d",
+		CurrentOriginBucket: "mainnetOrigin60m",
+	}
+	// Test HTTP client
+	testClient = httpclient.NewClient(httpclient.WithHTTPTimeout(5*time.Second), httpclient.WithRetryCount(0))
+	// Test errors
+	ErrResponseNotOK = errors.New("Response not OK")
+
 	// Selection of relays to populate the test InfluxDB (selected from this slice inside populateInfluxRelays)
 	testRelays = []struct {
 		applicationPublicKey,
@@ -105,6 +153,14 @@ var (
 )
 
 const (
+	// Test consts
+	phdBaseURL        = "http://localhost:8090"
+	apiKey            = "test_api_key_6789"
+	relayMeterBaseURL = "http://localhost:9898"
+	testUserID        = "12345678fgte0db3b6c63124"
+	mainBucket        = "mainnetRelay"
+	main1mBucket      = "mainnetRelayApp1m"
+
 	// Influx Tasks
 	app1mStringRaw = `from(bucket: "%s")
 	|> range(start: %s, stop: %s)
@@ -214,3 +270,245 @@ const (
 		"Applications": null
 	}`
 )
+
+/* Test Utility Funcs */
+
+// Initializes the Influx client and returns it alongside the org ID
+func initTestInfluxClient(c *require.Assertions) (influxdb2.Client, string) {
+	client := influxdb2.NewClientWithOptions(
+		influxOptions.URL, influxOptions.Token,
+		influxdb2.DefaultOptions().SetBatchSize(4_000),
+	)
+
+	dOrg, err := client.OrganizationsAPI().FindOrganizationByName(ctx, influxOptions.Org)
+	c.NoError(err)
+
+	health, err := client.Health(ctx)
+	c.NoError(err)
+	c.Equal("ready for queries and writes", *health.Message)
+
+	return client, *dOrg.Id
+}
+
+// Resets the Influx bucket each time the collector test runs
+func resetInfluxBuckets(client influxdb2.Client, orgID string, c *require.Assertions) {
+	bucketsAPI := client.BucketsAPI()
+	usedBuckets := []string{mainBucket, main1mBucket, influxOptions.CurrentBucket, influxOptions.CurrentOriginBucket}
+
+	for _, bucketName := range usedBuckets {
+		bucket, err := bucketsAPI.FindBucketByName(ctx, bucketName)
+		if err == nil {
+			err = client.BucketsAPI().DeleteBucketWithID(ctx, *bucket.Id)
+			c.NoError(err)
+		}
+		_, err = client.BucketsAPI().CreateBucketWithNameWithID(ctx, orgID, bucketName)
+		c.NoError(err)
+	}
+}
+
+// Initializes the Influx tasks used to populate each bucket from the main bucket
+func initInfluxTasks(tasksAPI influxAPI.TasksAPI, orgID string, c *require.Assertions) []*influxDomain.Task {
+	app1mString := fmt.Sprintf(app1mStringRaw, mainBucket, todayFormatted, endOfDayFormatted, main1mBucket)
+	app10mString := fmt.Sprintf(app10mStringRaw, main1mBucket, todayFormatted, endOfDayFormatted, influxOptions.CurrentBucket)
+	origin60mString := fmt.Sprintf(origin60mStringRaw, mainBucket, todayFormatted, endOfDayFormatted, influxOptions.CurrentOriginBucket)
+	taskNames := []string{"app-1m", "app-10m", "origin-sample-60m"}
+
+	existingTasks, err := tasksAPI.FindTasks(ctx, nil)
+	c.NoError(err)
+
+	tasks := []*influxDomain.Task{}
+	for i, fluxTask := range []string{app1mString, app10mString, origin60mString} {
+		taskName := taskNames[i]
+
+		for _, existingTask := range existingTasks {
+			if existingTask.Name == taskName {
+				err = tasksAPI.DeleteTaskWithID(ctx, existingTask.Id)
+				c.NoError(err)
+			}
+		}
+
+		task, err := tasksAPI.CreateTaskWithEvery(ctx, taskNames[i], fluxTask, "1h", orgID)
+		c.NoError(err)
+
+		tasks = append(tasks, task)
+	}
+
+	return tasks
+}
+
+// Saves a test batch of relays to the InfluxDB mainBucket
+func populateInfluxRelays(writeAPI influxAPI.WriteAPI, date time.Time, numberOfRelays int) {
+	timestampInterval := (24 * time.Hour) / time.Duration(numberOfRelays)
+
+	for i := 0; i < numberOfRelays; i++ {
+		index := i % 10
+		relay := testRelays[index]
+
+		relayTimestamp := date.Add(timestampInterval * time.Duration(i+1))
+		if i == 0 {
+			relayTimestamp.Add(time.Millisecond * 10)
+		}
+		if i == numberOfRelays-1 {
+			relayTimestamp.Add(-time.Millisecond * 10)
+		}
+
+		var result string
+		if i%9 != 0 {
+			result = "200"
+		} else {
+			result = "500"
+		}
+
+		relayPoint := influxdb2.NewPoint(
+			"relay",
+			map[string]string{
+				"applicationPublicKey": relay.applicationPublicKey,
+				"nodePublicKey":        relay.nodePublicKey,
+				"method":               relay.method,
+				"result":               result,
+				"blockchain":           relay.blockchain,
+				"blockchainSubdomain":  relay.blockchainSubdomain,
+				"host":                 "test_0bc93fa",
+				"region":               "us-east-2",
+			},
+			map[string]interface{}{
+				"bytes":       12345,
+				"elapsedTime": relay.elapsedTime,
+			},
+			relayTimestamp,
+		)
+		writeAPI.WritePoint(relayPoint)
+
+		originPoint := influxdb2.NewPoint(
+			"origin",
+			map[string]string{
+				"applicationPublicKey": relay.applicationPublicKey,
+			},
+			map[string]interface{}{
+				"origin": relay.origin,
+			},
+			relayTimestamp,
+		)
+		writeAPI.WritePoint(originPoint)
+	}
+
+	writeAPI.Flush()
+}
+
+// Initializes Pocket HTTP DB with required apps and LBs (will not recreate if they already exist)
+func populatePocketHTTPDB() error {
+	existingApps, err := get[[]repository.Application](phdBaseURL, "application", "", "")
+	if err != nil {
+		return err
+	}
+	existingLBs, err := get[[]repository.LoadBalancer](phdBaseURL, "load_balancer", "", "")
+	if err != nil {
+		return err
+	}
+	existingAppNames := []string{}
+	for _, app := range existingApps {
+		existingAppNames = append(existingAppNames, app.Name)
+	}
+	existingLBNames := []string{}
+	for _, lb := range existingLBs {
+		existingLBNames = append(existingLBNames, lb.Name)
+	}
+
+	for i, application := range testRelays {
+		/* Create Application -> POST /application */
+		var createdAppID string
+		if !stringUtils.ExactContains(existingAppNames, fmt.Sprintf("test-application-%d", i+1)) {
+			appInput := fmt.Sprintf(applicationJSON, i+1, testUserID, application.applicationPublicKey)
+			createdApplication, err := post[repository.Application](phdBaseURL, "application", []byte(appInput))
+			if err != nil {
+				return err
+			}
+			createdAppID = createdApplication.ID
+		}
+
+		/* Create Load Balancer -> POST /load_balancer */
+		if !stringUtils.ExactContains(existingLBNames, fmt.Sprintf("test-load-balancer-%d", i+1)) {
+			loadBalancerInput := fmt.Sprintf(loadBalancerJSON, i+1, testUserID, createdAppID)
+			_, err = post[repository.LoadBalancer](phdBaseURL, "load_balancer", []byte(loadBalancerInput))
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+// POST test util func
+func post[T any](baseURL, path string, postData []byte) (T, error) {
+	var data T
+
+	rawURL := fmt.Sprintf("%s/%s", baseURL, path)
+
+	headers := http.Header{
+		"Authorization": {apiKey},
+		"Content-Type":  {"application/json"},
+		"Connection":    {"Close"},
+	}
+
+	postBody := bytes.NewBufferString(string(postData))
+
+	response, err := testClient.Post(rawURL, postBody, headers)
+	if err != nil {
+		return data, err
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode != http.StatusOK {
+		return data, fmt.Errorf("%w. %s", ErrResponseNotOK, http.StatusText(response.StatusCode))
+	}
+
+	body, err := io.ReadAll(response.Body)
+	if err != nil {
+		return data, err
+	}
+
+	err = json.Unmarshal(body, &data)
+	if err != nil {
+		return data, err
+	}
+
+	return data, nil
+}
+
+// GET test util func
+func get[T any](baseURL, path, id, params string) (T, error) {
+	rawURL := fmt.Sprintf("%s/%s", baseURL, path)
+	if id != "" {
+		rawURL = fmt.Sprintf("%s/%s", rawURL, id)
+	}
+	if params != "" {
+		rawURL = fmt.Sprintf("%s%s", rawURL, params)
+	}
+
+	headers := http.Header{"Authorization": {apiKey}}
+
+	var data T
+
+	response, err := testClient.Get(rawURL, headers)
+	if err != nil {
+		return data, err
+	}
+	defer response.Body.Close()
+
+	if response.StatusCode != http.StatusOK {
+		return data, fmt.Errorf("%w. %s", ErrResponseNotOK, http.StatusText(response.StatusCode))
+	}
+
+	body, err := io.ReadAll(response.Body)
+	if err != nil {
+		return data, err
+	}
+
+	err = json.Unmarshal(body, &data)
+	if err != nil {
+		return data, err
+	}
+
+	return data, nil
+}
